handle string value

+++ cdc/sink/codec/avro_test.go
@@ -567,7 +567,7 @@ var avroTestColumns = []*avroTestColumnTuple{
 			Ft:            types.NewFieldType(mysql.TypeYear),
 		},
 		avroSchema{Type: "int", Parameters: map[string]string{"tidb_type": "YEAR"}},
-		int64(1970), "int",
+		int32(1970), "int",
 	},
 }
